@@ -1,12 +1,11 @@
-namespace RagePhoto.NET
-{
-    internal static class Program
+namespace RagePhoto.NET;
+
+internal static class Program {
+
+    [STAThread]
+    private static void Main()
     {
-        [STAThread]
-        private static void Main()
-        {
-            ApplicationConfiguration.Initialize();
-            Application.Run(new UserInterface());
-        }
+        ApplicationConfiguration.Initialize();
+        Application.Run(new UserInterface());
     }
 }
