@@ -1,55 +1,36 @@
-namespace RagePhoto.NET
-{
-    public partial class UserInterface : Form
-    {
-        public UserInterface()
-        {
-            InitializeComponent();
-        }
+using Syping.RagePhoto;
+namespace RagePhoto.NET;
 
-        private void UserInterface_Load(object sender, EventArgs e)
-        {
-        }
+public partial class UserInterface : Form {
 
-        private void OpenButton_Click(object sender, EventArgs e)
-        {
-            OpenFileDialog openFileDialog = new()
-            {
-                Title = "Open Photo...",
-                Filter = "RagePhoto compatible|PGTA5*;PRDR3*",
-                CheckFileExists = true,
-                CheckPathExists = true
-            };
-            openFileDialog.ShowDialog(this);
+    public UserInterface() {
+        InitializeComponent();
+    }
 
-            if (openFileDialog.FileName == "")
-                return;
+    private void OpenButton_Click(object sender, EventArgs e) {
+        OpenFileDialog openFileDialog = new() {
+            Title = "Open Photo...",
+            Filter = "RagePhoto compatible|PGTA5*;PRDR3*",
+            CheckFileExists = true,
+            CheckPathExists = true
+        };
 
-            RagePhoto ragePhoto = new();
-            bool isLoaded = ragePhoto.LoadFile(openFileDialog.FileName);
-            if (!isLoaded)
-            {
-                Int32 error = ragePhoto.GetError();
-                if (error <= (Int32)RagePhoto.Error.PhotoReadError)
-                {
-                    MessageBox.Show("Failed to read photo: " + openFileDialog.FileName, "Open Photo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                    return;
-                }
-            }
+        if (openFileDialog.ShowDialog(this) != DialogResult.OK)
+            return;
 
-            Image jpeg = ragePhoto.JPEG_Image;
-            ImageBox.Image = jpeg;
+        using Photo ragePhoto = new();
+        ragePhoto.LoadFile(openFileDialog.FileName);
 
-            String title = ragePhoto.Title;
-            if (String.IsNullOrEmpty(title))
-                Text = "RagePhoto.NET Photo Viewer";
-            else
-                Text = "RagePhoto.NET Photo Viewer - " + title;
-        }
+        using MemoryStream jpegStream = new(ragePhoto.Jpeg);
+        Image jpeg = Image.FromStream(jpegStream);
+        ImageBox.Image = jpeg;
 
-        private void CloseButton_Click(object sender, EventArgs e)
-        {
-            Close();
-        }
+        string title = ragePhoto.Title;
+        Text = string.IsNullOrEmpty(title) ?
+           "RagePhoto.NET Photo Viewer" : $"RagePhoto.NET Photo Viewer - {title}";
+    }
+
+    private void CloseButton_Click(object sender, EventArgs e) {
+        Close();
     }
 }
